refactored methods out from processSearchResultsActions file to dedicated files for each method
