+++ api/controllers/sessions/processSearchResults/processSearchResultsActions/failedRes.js
@@ -1,0 +1,7 @@
+exports.failed_res = (status, message) => {
+    return {
+        status: status,
+        success: false,
+        message: message
+    }
+}
